Fixed StateRelativeToState CohortDefinition and Evaluator.  Made the definition more flexible for more use cases.
DETAILS--
Renamed 'secondState' (ex. Died) to 'primaryState' (uses report dates) and firstState (ex. On ARVs) to relativeState (relative to primary).

Add variable to toggle between 'Start of State' and 'End of State' which uses the corresponding DAO methods (HavingState) instead of using 'In State'.

Removed the use of 'In State'.  Right now, only can choose between 'Start of State' and 'End of State'.  'In State' was going to be very complex so I will only implement it if it is needed by the report.

Add variable to toggle before and after, right now primaryState (Died) must come after relativeState (On ARVs).